+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]## [0.1.2.3b6]
 ### Added
 - DashPlotlyManager: support for scenario kpi compare
+- Visualization page KPIComparePage to automatically compare KPIs
 
 ## [0.1.2.3b5] - 2024-04-15
 ### Added